--- pyproject.toml
@@ -1,25 +1,0 @@
-[build-system]
-requires = ["setuptools","packaging", "setuptools-scm[toml]"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "minilayer"
-dynamic = ["dependencies", "optional-dependencies", "version"]
-readme = "README.md"
-
-[tool.setuptools.packages.find]
-include = ["minilayer*"]  
-
-[tool.setuptools.package-data]
-minilayer = ["py.typed"]
-
-[tool.setuptools.dynamic]
-dependencies = {file = ["requirements.txt"]}
-optional-dependencies={dev = {file = ["requirements-dev.txt"]},generate = {file = ["requirements-generate.txt"]}}
-
-
-[tool.setuptools_scm]
-
-[project.scripts]
-minilayer = "minilayer.__main__:main"
-
