--- tests/resources/t
@@ -1,1 +1,0 @@
-
