+++ tests/resources/test_feature_definitions/gh-cli/feature-definition.json
@@ -1,0 +1,49 @@
+{
+    "id": "gh-cli",
+    "version": "1.0.0",
+    "name": "GitHub CLI (via GitHub Releases)",
+    "documentationURL": "http://github.com/devcontainers-contrib/features/tree/main/src/gh-cli",
+    "description": "GitHub CLI is a command-line interface to GitHub for use in your terminal or your scripts.",
+    "options": {
+        "version": {
+            "default": "latest",
+            "description": "Select the version to install.",
+            "proposals": [
+                "latest"
+            ],
+            "type": "string"
+        }
+    },
+    "installsAfter": [
+        "ghcr.io/devcontainers-contrib/features/gh-release:1.0.0"
+    ],
+    "dependencies": [
+        {
+            "feature": "ghcr.io/devcontainers-contrib/features/gh-release:1.0.0",
+            "options": {
+                "repo": "cli/cli",
+                "target": "gh",
+                "version": "$options.version"
+            }
+        }
+    ],
+    "install_command": "echo 'Done!'",
+    "test_scenarios": [
+        {
+            "name": "test_defaults_debian",
+            "image": "mcr.microsoft.com/devcontainers/base:debian",
+            "test_commands": [
+                "gh --version"
+            ],
+            "options": {}
+        },
+        {
+            "name": "test_defaults_alpine",
+            "image": "mcr.microsoft.com/devcontainers/base:alpine",
+            "test_commands": [
+                "gh --version"
+            ],
+            "options": {}
+        }
+    ]
+}
